use core ViewCreator and do not fail when no default views exist
